refactor: consolidate public intake flow and deduplicate practice config resolution
